refactor: LottoCount 객체 도입
LottoCount 객체를 도입하면서
LottosFactory 가 하는 일을 줄이고,
LottosFactory 에서 제공하던 기능을
Lottos 로 옮김.

+++ src/main/java/lotto/view/InputView.java
@@ -36,7 +36,7 @@ public class InputView {
 	}
 
 
-	public static List<List<Integer>> getManualLottos(int manualLottoAmount) {
+	public static List<List<Integer>> getManualLottos(long manualLottoAmount) {
 		List<List<Integer>> lottos = new ArrayList<>();
 
 		OutputView.askEnterManualLottos();
